requirements.txt
django
django-tailwind
django-tailwind[reload]
python-dotenv
ruff
djlint
pre-commit
mkdocs-material
gunicorn
whitenoise
social-auth-app-django
tzlocal